Initial commit: ROOT6 SB, README.md, and LICENSE
Here we go (for the first time). Let's collect what is necessary
for bachelor and master students doing particle physics in one
place, first of all: ROOT6. The root6.SlackBuild works fine on my
Slackware64-current, but of course it needs to be tested at least
on i686 -current (or ARM if it is possible).